+++ CHANGELOG.md
@@ -1,4 +1,9 @@
-# 0.4.0 (10-04-2024)
+# 0.5.0 (27-04-2024)
+
+- **added:** `initialDelay` and `delayIncrement` options in `retryAsync` function. See [usage of retryAsync](https://www.npmjs.com/package/utility-kit#utility)
+- **removed:** `showError` option in favour of `onError` callback in `retry` and `retryAsync` functions. See [usage of retry and retryAsync](https://www.npmjs.com/package/utility-kit#utility)
+
+## 0.4.0 (10-04-2024)
 
 - **added:** `random` function. See [usage of random](https://www.npmjs.com/package/utility-kit#random)